misc changes
