+++ .eslintrc.yml
@@ -1,0 +1,27 @@
+env:
+  browser: true
+  es6: true
+  node: true
+  jest: true
+extends:
+  - 'eslint:recommended'
+parserOptions:
+  ecmaFeatures:
+    experimentalObjectRestSpread: true
+    jsx: true
+  sourceType: module
+rules:
+  indent:
+    - error
+    - 2
+  linebreak-style:
+    - error
+    - unix
+  quotes:
+    - error
+    - single
+  semi:
+    - error
+    - always
+  no-console:
+    - warn
