+++ app/app.js
@@ -1,3 +1,24 @@
+/**
+ * Copyright 2016 IBM Corp. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the “License”);
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an “AS IS” BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*globals $:false */
+/*global Bloodhound:true*/
+/*global traversal:true*/
+/*global vis:true*/
+
 // Find the Bacon number
 var ajaxRequest = [];
 $(document).ready(function () {
@@ -25,82 +46,78 @@ $(document).ready(function () {
       inProgressContainer.addClass('hidden');
     }
     var value = $(this).val();
-    callService(value, "", "");
+    callService(value, '', '');
 
   });
 
-  $("#the-result").click( function()
-       {
-         var inProgressContainer = $('#the-result').parent().parent();
-         if (!inProgressContainer.hasClass('hidden')) {
-           inProgressContainer.addClass('hidden');
-         }
+  $('#the-result').click( function() {
+    var inProgressContainer = $('#the-result').parent().parent();
+    if (!inProgressContainer.hasClass('hidden')) {
+      inProgressContainer.addClass('hidden');
+    }
+  });
 
-       }
-    );
+  $('a[href="#sync"]').click(function() {
+    var graphVisContainer = $('#the-graph').parent().parent();
+    if (!graphVisContainer.hasClass('hidden')) {
+      graphVisContainer.addClass('hidden');
+    }
+    callSync();
+  });
 
-    $('a[href="#sync"]').click(function(){
-      var graphVisContainer = $('#the-graph').parent().parent();
-      if (!graphVisContainer.hasClass('hidden')) {
-
-          graphVisContainer.addClass('hidden');
-        }
-      callSync();
-    });
-
-    $('a[href="#reset"]').click(function(){
-      var graphVisContainer = $('#the-graph').parent().parent();
-      if (!graphVisContainer.hasClass('hidden')) {
-          graphVisContainer.addClass('hidden');
-        }
-      callReset();
-    });
+  $('a[href="#reset"]').click(function() {
+    var graphVisContainer = $('#the-graph').parent().parent();
+    if (!graphVisContainer.hasClass('hidden')) {
+      graphVisContainer.addClass('hidden');
+    }
+    callReset();
+  });
 
 });
+
 function callSync()
 {
   var inProgressContainer = $('#the-result').parent().parent();
   var loadingSpinner = '<i class="fa fa-spinner fa-spin"></i>';
   $('#the-result').html(loadingSpinner);
   if (inProgressContainer.hasClass('hidden')) {
-      inProgressContainer.removeClass('hidden');
-    }
-    var url = '/knowledgegraph/syncService';
-    ajaxRequest.push($.get(url, function (data) {
-      $('#the-result').html('<P>'+data+'</P><button id="inprogress" >Click here to close</button>');
-    }));
-
+    inProgressContainer.removeClass('hidden');
+  }
+  var url = '/knowledgegraph/syncService';
+  ajaxRequest.push($.get(url, function (data) {
+    $('#the-result').html('<P>'+data+'</P><button id="inprogress" >Click here to close</button>');
+  }));
 }
+
 function callReset()
 {
   var inProgressContainer = $('#the-result').parent().parent();
   var loadingSpinner = '<i class="fa fa-spinner fa-spin"></i>';
   $('#the-result').html(loadingSpinner);
   if (inProgressContainer.hasClass('hidden')) {
-      inProgressContainer.removeClass('hidden');
-    }
-    var url = '/knowledgegraph/resetService';
-    ajaxRequest.push($.get(url, function (data) {
-        $('#the-result').html('<P>'+data+'</P><button id="inprogress" type="button">Click here to close</button>');
-    }));
+    inProgressContainer.removeClass('hidden');
+  }
+  var url = '/knowledgegraph/resetService';
+  ajaxRequest.push($.get(url, function (data) {
+    $('#the-result').html('<P>'+data+'</P><button id="inprogress" type="button">Click here to close</button>');
+  }));
 }
+
 function callService(value, nodeLabel, nodeText){
   var graphVisContainer = $('#the-graph').parent().parent();
-  var queryContainer = $('#query').parent().parent().parent().parent();
+  // var queryContainer = $('#query').parent().parent().parent().parent();
 
   // Loading spinner
   var loadingSpinner = '<i class="fa fa-spinner fa-spin"></i>';
   $('#query, #the-graph').html(loadingSpinner);
 
-
   var url = '/knowledgegraph/' + value;
-  if(nodeLabel && (nodeLabel!=""))
+  if(nodeLabel && (nodeLabel!=''))
   {
     url = url+ '/' + nodeLabel;
   }
 
-
-  if(nodeText && (nodeText!=""))
+  if(nodeText && (nodeText!=''))
   {
     url = url+ '/' + nodeText;
   }
@@ -115,15 +132,13 @@ function callService(value, nodeLabel, nodeText){
     traversal.allPaths();
   }
 
-
-
   if (ajaxRequest.length > 0) {
     ajaxRequest[0].abort();
   }
 
   if (graphVisContainer.hasClass('hidden')) {
-      graphVisContainer.removeClass('hidden');
-    }
+    graphVisContainer.removeClass('hidden');
+  }
 
   ajaxRequest.push($.get(url, function (data) {
     console.log(data);
@@ -135,10 +150,9 @@ function callService(value, nodeLabel, nodeText){
     var rawEdges = [];
     var ignoreEdges = [];
 
-
-    for (i = 0; i < data.data.length; i++) {
+    for (let i = 0; i < data.data.length; i++) {
       var path = data.data[i].objects;
-      for (j = 0; j < path.length; j++) {
+      for (let j = 0; j < path.length; j++) {
         var obj = path[j];
         if (obj.type == 'vertex') {
           if (ignoreNodes.indexOf(obj.id) < 0) {
@@ -166,8 +180,7 @@ function callService(value, nodeLabel, nodeText){
                 border:'black',
               },
             };
-            if(obj.properties.sentiment)
-            {
+            if(obj.properties.sentiment) {
               if (obj.properties.sentiment[0].value == 'Neutral') {
                 nodeObject.color = {
                   background:'#048BF9',
@@ -178,7 +191,7 @@ function callService(value, nodeLabel, nodeText){
                     border:'black',
                   },
                 };
-              }else if (obj.properties.sentiment[0].value == 'Positive') {
+              } else if (obj.properties.sentiment[0].value == 'Positive') {
                 nodeObject.color = {
                   background:'#2B991B',
                   border:'#713E7F',
@@ -188,7 +201,7 @@ function callService(value, nodeLabel, nodeText){
                     border:'black',
                   },
                 };
-              }else if (obj.properties.sentiment[0].value == 'Negative') {
+              } else if (obj.properties.sentiment[0].value == 'Negative') {
                 nodeObject.color = {
                   background:'#F90404',
                   border:'#713E7F',
@@ -200,13 +213,10 @@ function callService(value, nodeLabel, nodeText){
                 };
 
               }
-
             }
             rawNodes.push(nodeObject);
             ignoreNodes.push(obj.id);
           }
-
-
         }
 
         if (obj.type == 'edge') {
@@ -222,7 +232,6 @@ function callService(value, nodeLabel, nodeText){
           }
         }
       }
-
     }
 
     console.log(rawNodes);
@@ -234,45 +243,45 @@ function callService(value, nodeLabel, nodeText){
     var edges = new vis.DataSet(rawEdges);
 
     // create a network
+    /*eslint no-unused-vars: ["error", { "varsIgnorePattern": "container" || "options" }]*/
     var container = document.getElementById('the-graph');
-    var data = {
+    data = {
       nodes: nodes,
       edges: edges,
     };
-    var layoutMethod = "directed";
+    var layoutMethod = 'directed';
     var options = {
-        layout: {
-          hierarchical: {
-            sortMethod: layoutMethod,
-            levelSeparation: 200
-          }
-        },
-        edges: {
-          smooth: true,
-          arrows: {to : true },
-          forceDirection: "none"
-        },
-        interaction:{
-          dragNodes:false,
-          dragView: false,
-          hideEdgesOnDrag: false,
-          hideNodesOnDrag: false,
-          hover: true,
-          hoverConnectedEdges: true,
-          keyboard: {
-            enabled: false,
-            speed: {x: 10, y: 10, zoom: 0.02},
-            bindToWindow: true
-          },
-          multiselect: true,
-          navigationButtons: true,
-          selectable: true,
-          selectConnectedEdges: true,
-          tooltipDelay: 300,
-          zoomView: true
+      layout: {
+        hierarchical: {
+          sortMethod: layoutMethod,
+          levelSeparation: 200
         }
-      };
-
+      },
+      edges: {
+        smooth: true,
+        arrows: {to : true },
+        forceDirection: 'none'
+      },
+      interaction:{
+        dragNodes:false,
+        dragView: false,
+        hideEdgesOnDrag: false,
+        hideNodesOnDrag: false,
+        hover: true,
+        hoverConnectedEdges: true,
+        keyboard: {
+          enabled: false,
+          speed: {x: 10, y: 10, zoom: 0.02},
+          bindToWindow: true
+        },
+        multiselect: true,
+        navigationButtons: true,
+        selectable: true,
+        selectConnectedEdges: true,
+        tooltipDelay: 300,
+        zoomView: true
+      }
+    };
 
     var network = new vis.Network(container, data, options);
