+++ lib/JanusGraphClient.js
@@ -16,41 +16,42 @@
  * limitations under the License.
  */
 
+/* eslint-disable indent */
+/* eslint-disable  no-useless-escape */
+
 const rp = require('request-promise');
 
 class JanusGraphClient {
 
-    /**
-     * Creates a new instance of JanusGraphClient.
-     * @param {String} url - JanusGraph url
-     * @param {String} username - JanusGraph username
+	/**
+	 * Creates a new instance of JanusGraphClient.
+	 * @param {String} url - JanusGraph url
+	 * @param {String} username - JanusGraph username
 	 * @param {String} password - JanusGraph password
-     */
-    constructor(graphurl, username, password) {
-    	this.url = graphurl;
-    	this.username = username;
-    	this.password = password;
-    	this.authHeader = 'Basic '+ new Buffer(this.username + ":" + this.password).toString("base64");
-    }
+	 */
+	constructor(graphurl, username, password) {
+    this.url = graphurl;
+    this.username = username;
+    this.password = password;
+    this.authHeader = 'Basic '+ new Buffer(this.username + ':' + this.password).toString('base64');
+  }
 
-    getPostOptions(body) {
-    	let options = {
+	getPostOptions(body) {
+		let options = {
 			method: 'POST',
 			uri: this.url,
 			headers: {
-				"Authorization": this.authHeader,
-				"Content-Type": "application/json"
+				'Authorization': this.authHeader,
+				'Content-Type': 'application/json'
 			},
 			json: true,
 			body: body
 		};
-
-
-    	return options;
+		return options;
 	}
 
 	graphExists(id) {
-    	console.log(`Checking if graph exists with id '${id}'...`);
+		console.log(`Checking if graph exists with id '${id}'...`);
 		const body = {
 			gremlin: `def graph=ConfiguredGraphFactory.open("${id}");0;`
 		};
@@ -93,7 +94,7 @@ class JanusGraphClient {
 					return Promise.resolve(responseBody);
 				}
 				else {
-                    console.log("Error executing the query.."+responseBody);
+					console.log('Error executing the query..' + responseBody);
 					return Promise.reject(new Error('Invalid status returned from server. Actual Error:'+responseBody));
 				}
 			});
@@ -105,10 +106,10 @@ class JanusGraphClient {
 			if (vertex.hasOwnProperty(property) && property != 'label') {
 				let value = vertex[property];
 				if (this.isStringValue(value)) {
-					gremlinQuery += `, "${property}", "${this.escapeStringValue(value)}"`
+					gremlinQuery += `, "${property}", "${this.escapeStringValue(value)}"`;
 				}
 				else {
-					gremlinQuery += `, "${property}", ${value}`
+					gremlinQuery += `, "${property}", ${value}`;
 				}
 
 			}
@@ -134,10 +135,10 @@ class JanusGraphClient {
 			if (properties.hasOwnProperty(property)) {
 				let value = properties[property];
 				if (this.isStringValue(value)) {
-					gremlinQuery += `, "${property}", "${this.escapeStringValue(value)}"`
+					gremlinQuery += `, "${property}", "${this.escapeStringValue(value)}"`;
 				}
 				else {
-					gremlinQuery += `, "${property}", ${value}`
+					gremlinQuery += `, "${property}", ${value}`;
 				}
 			}
 		}
@@ -160,10 +161,10 @@ class JanusGraphClient {
 			if (properties.hasOwnProperty(property)) {
 				let value = properties[property];
 				if (this.isStringValue(value)) {
-					gremlinQuery += `.property("${property}", "${this.escapeStringValue(value)}")`
+					gremlinQuery += `.property("${property}", "${this.escapeStringValue(value)}")`;
 				}
 				else {
-					gremlinQuery += `.property("${property}", ${value})`
+					gremlinQuery += `.property("${property}", ${value})`;
 				}
 			}
 		}
@@ -172,7 +173,7 @@ class JanusGraphClient {
 	}
 
 	isStringValue(value) {
-    	return Object.prototype.toString.call(value) == '[object String]';
+		return Object.prototype.toString.call(value) == '[object String]';
 	}
 
 	escapeStringValue(value) {
