+++ templates/index.jade
@@ -10,8 +10,8 @@ head
     link(href="//raw.githubusercontent.com/bassjobsen/typeahead.js-bootstrap-css/master/typeaheadjs.css", rel="stylesheet")
     link(href="//cdnjs.cloudflare.com/ajax/libs/vis/4.5.0/vis.min.css", rel="stylesheet")
     link(rel="stylesheet", href="//maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css")
-    link(rel="stylesheet", href="/assets/master.css")
-    link(rel="stylesheet", href="/assets/style.css")
+    link(rel="stylesheet", href="/public/master.css")
+    link(rel="stylesheet", href="/public/style.css")
     // HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries
     //if lt IE 9
       script(src='//oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js')
@@ -125,5 +125,5 @@ body
   script(src="//maxcdn.bootstrapcdn.com/bootstrap/3.3.5/js/bootstrap.min.js")
   script(src="//cdnjs.cloudflare.com/ajax/libs/typeahead.js/0.11.1/typeahead.bundle.min.js")
   script(src="//cdnjs.cloudflare.com/ajax/libs/vis/4.5.0/vis.min.js")
-  script(type="text/javascript", src="assets/traversal.js")
-  script(type="text/javascript", src="assets/app.js")
+  script(type="text/javascript", src="app/traversal.js")
+  script(type="text/javascript", src="app/app.js")
