+++ tools/updateConfiguration.py
